examples/code_review_agent: split Approver into its own file
Move Approver and its support types out of governance.go into approver.go.

Adjust the interactive input unavailable subtest.

Updates #2004
